# ext/bcmath: Supports scientific notation