--- a/ext/bcmath/libbcmath/src/str2num.c
+++ b/ext/bcmath/libbcmath/src/str2num.c
@@ -109,7 +109,7 @@ static bool bc_scientific_notation_str2num(
 	bc_num *num, const char *str, const char *end, const char *integer_ptr, const char *fractional_ptr, const char *exponent_ptr,
 	size_t digits, size_t *full_scale)
 {
-	const char *fractional_end = exponent_ptr;
+	const char *fractional_end = fractional_ptr != NULL ? exponent_ptr : NULL;
 
 	if (UNEXPECTED(*exponent_ptr != 'e' && *exponent_ptr != 'E')) {
 		goto fail;
@@ -173,7 +173,9 @@ static bool bc_scientific_notation_str2num(
 	}
 
 	nptr = bc_copy_and_toggle_bcd(nptr, integer_ptr, integer_end);
-	nptr = bc_copy_and_toggle_bcd(nptr, fractional_ptr, fractional_end);
+	if (fractional_ptr != NULL) {
+		nptr = bc_copy_and_toggle_bcd(nptr, fractional_ptr, fractional_end);
+	}
 
 	if (digits > str_full_len) {
 		/* Fill the rest integer part with zeros */
@@ -221,8 +223,8 @@ bool bc_str2num(bc_num *num, const char *str, const char *end, size_t scale, siz
 	const char *decimal_point = (*ptr == '.') ? ptr : NULL;
 
 	/* If a non-digit and non-decimal-point indicator is in the string, i.e. an invalid character */
-	if (UNEXPECTED(!decimal_point && *ptr != '\0')) {
-		goto fail;
+	if (!decimal_point && *ptr != '\0') {
+		return bc_scientific_notation_str2num(num, str, end, integer_ptr, fractional_ptr, ptr, digits, full_scale);
 	}
 
 	/* search and validate fractional end if exists */
@@ -311,8 +313,4 @@ after_fractional:
 zero:
 	*num = bc_copy_num(BCG(_zero_));
 	return true;
-
-fail:
-	*num = bc_copy_num(BCG(_zero_));
-	return false;
 }