@j3j5 2026-08-18
This is super useful, is there any reason why it wasn't merged? is there anything that could be done to help getting it merged? I see on [the mailing list](https://externals.io/message/126775) there wasn't really any objection, but it got stuck.